Move ForAll to dispatcher
