+++ numba/cuda/dispatcher.py
@@ -16,7 +16,7 @@ from numba.core.typing.typeof import Purpose, typeof
 
 from numba.cuda.api import get_current_device
 from numba.cuda.args import wrap_arg
-from numba.cuda.compiler import compile_cuda, ForAll
+from numba.cuda.compiler import compile_cuda
 from numba.cuda.cudadrv import driver
 from numba.cuda.cudadrv.devices import get_context
 from numba.cuda.cudadrv.libs import get_cudalib
@@ -375,6 +375,51 @@ class _Kernel(serialize.ReduceMixin):
 
         else:
             raise NotImplementedError(ty, val)
+
+
+class ForAll(object):
+    def __init__(self, kernel, ntasks, tpb, stream, sharedmem):
+        if ntasks < 0:
+            raise ValueError("Can't create ForAll with negative task count: %s"
+                             % ntasks)
+        self.kernel = kernel
+        self.ntasks = ntasks
+        self.thread_per_block = tpb
+        self.stream = stream
+        self.sharedmem = sharedmem
+
+    def __call__(self, *args):
+        if self.ntasks == 0:
+            return
+
+        if self.kernel.specialized:
+            kernel = self.kernel
+        else:
+            kernel = self.kernel.specialize(*args)
+        blockdim = self._compute_thread_per_block(kernel)
+        griddim = (self.ntasks + blockdim - 1) // blockdim
+
+        return kernel[griddim, blockdim, self.stream, self.sharedmem](*args)
+
+    def _compute_thread_per_block(self, kernel):
+        tpb = self.thread_per_block
+        # Prefer user-specified config
+        if tpb != 0:
+            return tpb
+        # Else, ask the driver to give a good config
+        else:
+            ctx = get_context()
+            # Kernel is specialized, so there's only one definition - get it so
+            # we can get the cufunc from the code library
+            defn = next(iter(kernel.overloads.values()))
+            kwargs = dict(
+                func=defn._codelibrary.get_cufunc(),
+                b2d_func=0,     # dynamic-shared memory is constant to blksz
+                memsize=self.sharedmem,
+                blocksizelimit=1024,
+            )
+            _, tpb = ctx.get_max_potential_block_size(**kwargs)
+            return tpb
 
 
 class _LaunchConfiguration:
